Rearranging npc-recharge macros.
